django-docker/requirements.txt
Django>=5.0,<6.0
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-extensions

argon2-cffi>=23.1.0
